requirements/requirements.txt
boto3>=1.26.0
PyJWT==2.6.0
cryptography>=3.0.0
uvicorn>=0.18.0